Move public headers into wlcs/ subdirectory
